+++ CHANGELOG.md
@@ -12,6 +12,8 @@ is based on https://keepachangelog.com/en/1.1.0/.
   specification.
 - A `types::branch::Map` now holds at most 32 branches. This is reflected in a
   changed value for `MAX_BRANCHES` and `raw_map` returning an `u32`.
+- `decoder::payload::AddressInfo::address` is now an `i64` which is always
+  sign-extended.
 
 ### Added